Fix MSVC warning.

+++ js/src/jsatom.h
@@ -294,7 +294,7 @@ struct AtomStateEntry {
     AtomStateEntry(const AtomStateEntry &other) : bits(other.bits) {}
 
     AtomStateEntry(JSFixedString *futureAtom, bool intern)
-      : bits(uintptr_t(futureAtom) | intern)
+      : bits(uintptr_t(futureAtom) | uintptr_t(intern))
     {}
 
     bool isInterned() const {
